Fix edge case
