+++ packages/graphql/src/translate/queryAST/factory/FilterFactory.ts
@@ -687,7 +687,7 @@ export class FilterFactory {
                     // This conditional handles when the relationship is an interface which is also being accessed through an interface
                     if (
                         isInterfaceEntity(relationship.target) &&
-                        Object.keys(value).includes(relationship.propertiesTypeName)
+                        Object.keys(value).some((v) => relationship.siblings?.includes(v))
                     ) {
                         return Object.entries(value).flatMap(([k, v]) => {
                             if (k === relationship.propertiesTypeName) {
